requirement.txt
Django==2.1.8
django-filter==2.2.0
djangorestframework==3.10.1
PyMySQL==0.9.3
pytz==2019.1
redis==3.2.1
